Containerized Flask/requirements.txt
flask==2.1.2
psycopg2-binary==2.9.3
psycopg==3.0.12
redis==4.2.2
yoyo-migrations==7.3.2
